requirements.txt
appdirs==1.4.4
bcrypt==3.1.7
blinker==1.4
certifi==2020.6.20
cffi==1.14.0
click==7.1.2
distlib==0.3.1
dnspython==2.0.0
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql==0.0.2
mysql-connector-python==8.0.21
mysqlclient==2.0.1
Pillow==7.2.0
pipenv==2020.6.2
protobuf==3.12.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
six==1.15.0
SQLAlchemy==1.3.18
virtualenv==20.0.27
virtualenv-clone==0.5.4
Werkzeug==1.0.1
WTForms==2.3.1
